Baza danych polaczona z wykresem, prawie dziala, przycisk - nie dziala

+++ app/src/main/java/com/example/telemedycynaapp/database/Meassure.kt
@@ -5,8 +5,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "meassures")
 data class Meassure(
+
     @PrimaryKey(autoGenerate = true)
-    val id: Long = 0,
+    val id: Int = 0,
     val date: String,
     val humidity: Float
 )
